Move results pane to separate class. Add copyright notices
